requirements.txt
torch
torchvision~=0.6.0
tensorboard
sacred
PyYAML~=5.3.1
numpy~=1.18.4
torchsummary~=1.5.1
pillow~=7.1.2